Upgrade upstream and add reset shared button

+++ index.bundle.js
@@ -509,8 +509,112 @@ function markdownToBlocks(md) {
 }
 
 },{"./Edit":1,"./docs":2,"array-shuffle":19,"curtiz-parse-markdown":48,"curtiz-quiz-planner":51,"curtiz-utils":52,"curtiz-web-db":54,"isomorphic-gatty":87,"jmdict-furigana-node":105,"react":173,"react-dom":170}],4:[function(require,module,exports){
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+class Store {
+    constructor(dbName = 'keyval-store', storeName = 'keyval') {
+        this.storeName = storeName;
+        this._dbName = dbName;
+        this._storeName = storeName;
+        this._init();
+    }
+    _init() {
+        if (this._dbp) {
+            return;
+        }
+        this._dbp = new Promise((resolve, reject) => {
+            const openreq = indexedDB.open(this._dbName, 1);
+            openreq.onerror = () => reject(openreq.error);
+            openreq.onsuccess = () => resolve(openreq.result);
+            // First time setup: create an empty object store
+            openreq.onupgradeneeded = () => {
+                openreq.result.createObjectStore(this._storeName);
+            };
+        });
+    }
+    _withIDBStore(type, callback) {
+        this._init();
+        return this._dbp.then(db => new Promise((resolve, reject) => {
+            const transaction = db.transaction(this.storeName, type);
+            transaction.oncomplete = () => resolve();
+            transaction.onabort = transaction.onerror = () => reject(transaction.error);
+            callback(transaction.objectStore(this.storeName));
+        }));
+    }
+    _close() {
+        this._init();
+        return this._dbp.then(db => {
+            db.close();
+            this._dbp = undefined;
+        });
+    }
+}
+let store;
+function getDefaultStore() {
+    if (!store)
+        store = new Store();
+    return store;
+}
+function get(key, store = getDefaultStore()) {
+    let req;
+    return store._withIDBStore('readwrite', store => {
+        req = store.get(key);
+    }).then(() => req.result);
+}
+function set(key, value, store = getDefaultStore()) {
+    return store._withIDBStore('readwrite', store => {
+        store.put(value, key);
+    });
+}
+function update(key, updater, store = getDefaultStore()) {
+    return store._withIDBStore('readwrite', store => {
+        const req = store.get(key);
+        req.onsuccess = () => {
+            store.put(updater(req.result), key);
+        };
+    });
+}
+function del(key, store = getDefaultStore()) {
+    return store._withIDBStore('readwrite', store => {
+        store.delete(key);
+    });
+}
+function clear(store = getDefaultStore()) {
+    return store._withIDBStore('readwrite', store => {
+        store.clear();
+    });
+}
+function keys(store = getDefaultStore()) {
+    const keys = [];
+    return store._withIDBStore('readwrite', store => {
+        // This would be store.getAllKeys(), but it isn't supported by Edge or Safari.
+        // And openKeyCursor isn't supported by Safari.
+        (store.openKeyCursor || store.openCursor).call(store).onsuccess = function () {
+            if (!this.result)
+                return;
+            keys.push(this.result.key);
+            this.result.continue();
+        };
+    }).then(() => keys);
+}
+function close(store = getDefaultStore()) {
+    return store._close();
+}
+
+exports.Store = Store;
+exports.get = get;
+exports.set = set;
+exports.update = update;
+exports.del = del;
+exports.clear = clear;
+exports.keys = keys;
+exports.close = close;
+
+},{}],5:[function(require,module,exports){
 const path = require("./path.js");
-const { ENOENT, EEXIST, ENOTEMPTY } = require("./errors.js");
+const { EEXIST, ENOENT, ENOTDIR, ENOTEMPTY } = require("./errors.js");
 
 const STAT = 0;
 
@@ -619,17 +723,16 @@ module.exports = class CacheFS {
   _lookup(filepath, follow = true) {
     let dir = this._root;
     let partialPath = '/'
-    for (let part of path.split(filepath)) {
+    let parts = path.split(filepath)
+    for (let i = 0; i < parts.length; ++ i) {
+      let part = parts[i];
       dir = dir.get(part);
       if (!dir) throw new ENOENT(filepath);
       // Follow symlinks
-      if (follow) {
+      if (follow || i < parts.length - 1) {
         const stat = dir.get(STAT)
         if (stat.type === 'symlink') {
-          let target = stat.target
-          if (!target.startsWith('/')) {
-            target = path.normalize(path.join(partialPath, target))
-          }
+          let target = path.resolve(partialPath, stat.target)
           dir = this._lookup(target)
         }
         if (!partialPath) {
@@ -660,8 +763,10 @@ module.exports = class CacheFS {
     dir.set(basename, entry);
   }
   rmdir(filepath) {
+    let dir = this._lookup(filepath);
+    if (dir.get(STAT).type !== 'dir') throw new ENOTDIR();
     // check it's empty (size should be 1 for just StatSym)
-    if (this._lookup(filepath).size > 1) throw new ENOTEMPTY();
+    if (dir.size > 1) throw new ENOTEMPTY();
     // remove from parent
     let parent = this._lookup(path.dirname(filepath));
     let basename = path.basename(filepath);
@@ -669,13 +774,14 @@ module.exports = class CacheFS {
   }
   readdir(filepath) {
     let dir = this._lookup(filepath);
+    if (dir.get(STAT).type !== 'dir') throw new ENOTDIR();
     return [...dir.keys()].filter(key => typeof key === "string");
   }
-  writeFile(filepath, data, { mode }) {
+  writeStat(filepath, size, { mode }) {
     let ino;
     try {
       let oldStat = this.stat(filepath);
-      if (mode === null) {
+      if (mode == null) {
         mode = oldStat.mode;
       }
       ino = oldStat.ino;
@@ -691,7 +797,7 @@ module.exports = class CacheFS {
     let stat = {
       mode,
       type: "file",
-      size: data.length,
+      size,
       mtimeMs: Date.now(),
       ino,
     };
@@ -759,13 +865,13 @@ module.exports = class CacheFS {
   }
 };
 
-},{"./errors.js":11,"./path.js":13}],5:[function(require,module,exports){
+},{"./errors.js":12,"./path.js":14}],6:[function(require,module,exports){
 module.exports = class HttpBackend {
   constructor(url) {
     this._url = url;
   }
   loadSuperblock() {
-    return fetch(this._url + '/.superblock.txt').then(res => res.text())
+    return fetch(this._url + '/.superblock.txt').then(res => res.ok ? res.text() : null)
   }
   async readFile(filepath) {
     const res = await fetch(this._url + filepath)
@@ -775,10 +881,18 @@ module.exports = class HttpBackend {
       throw new Error('ENOENT')
     }
   }
+  async sizeFile(filepath) {
+    const res = await fetch(this._url + filepath, { method: 'HEAD' })
+    if (res.status === 200) {
+      return res.headers.get('content-length')
+    } else {
+      throw new Error('ENOENT')
+    }
+  }
 }
 
-},{}],6:[function(require,module,exports){
-const idb = require("@wmhilton/idb-keyval");
+},{}],7:[function(require,module,exports){
+const idb = require("@isomorphic-git/idb-keyval");
 
 module.exports = class IdbBackend {
   constructor(name) {
@@ -808,8 +922,8 @@ module.exports = class IdbBackend {
   }
 }
 
-},{"@wmhilton/idb-keyval":14}],7:[function(require,module,exports){
-const idb = require("@wmhilton/idb-keyval");
+},{"@isomorphic-git/idb-keyval":4}],8:[function(require,module,exports){
+const idb = require("@isomorphic-git/idb-keyval");
 
 const sleep = ms => new Promise(r => setTimeout(r, ms))
 
@@ -904,7 +1018,7 @@ module.exports = class Mutex {
   }
 }
 
-},{"@wmhilton/idb-keyval":14}],8:[function(require,module,exports){
+},{"@isomorphic-git/idb-keyval":4}],9:[function(require,module,exports){
 const { encode, decode } = require("isomorphic-textencoder");
 const debounce = require("just-debounce-it");
 
@@ -940,7 +1054,7 @@ function cleanParams2(oldFilepath, newFilepath) {
 }
 
 module.exports = class PromisifiedFS {
-  constructor(name, { wipe, url } = {}) {
+  constructor(name, { wipe, url, urlauto } = {}) {
     this._name = name
     this._idb = new IdbBackend(name);
     this._mutex = new Mutex(name);
@@ -952,6 +1066,7 @@ module.exports = class PromisifiedFS {
     }, 500);
     if (url) {
       this._http = new HttpBackend(url)
+      this._urlauto = !!urlauto
     }
     this._operations = new Set()
 
@@ -966,6 +1081,7 @@ module.exports = class PromisifiedFS {
     this.lstat = this._wrap(this.lstat, false)
     this.readlink = this._wrap(this.readlink, false)
     this.symlink = this._wrap(this.symlink, true)
+    this.backFile = this._wrap(this.backFile, true)
 
     this._deactivationPromise = null
     this._deactivationTimeout = null
@@ -1050,18 +1166,46 @@ module.exports = class PromisifiedFS {
       await this._idb.saveSuperblock(this._cache._root);
     }
   }
+  async _writeStat(filepath, size, opts) {
+    let dirparts = path.split(path.dirname(filepath))
+    let dir = dirparts.shift()
+    for (let dirpart of dirparts) {
+      dir = path.join(dir, dirpart)
+      try {
+        this._cache.mkdir(dir, { mode: 0o777 })
+      } catch (e) {}
+    }
+    return this._cache.writeStat(filepath, size, opts)
+  }
   async readFile(filepath, opts) {
     ;[filepath, opts] = cleanParams(filepath, opts);
     const { encoding } = opts;
     if (encoding && encoding !== 'utf8') throw new Error('Only "utf8" encoding is supported in readFile');
-    const stat = this._cache.stat(filepath);
-    let data = await this._idb.readFile(stat.ino)
+    let data = null, stat = null
+    try {
+      stat = this._cache.stat(filepath);
+      data = await this._idb.readFile(stat.ino)
+    } catch (e) {
+      if (!this._urlauto) throw e
+    }
     if (!data && this._http) {
+      let lstat = this._cache.lstat(filepath)
+      while (lstat.type === 'symlink') {
+        filepath = path.resolve(path.dirname(filepath), lstat.target)
+        lstat = this._cache.lstat(filepath)
+      }
       data = await this._http.readFile(filepath)
     }
-    if (data && encoding === "utf8") {
-      data = decode(data);
+    if (data) {
+      if (!stat || stat.size != data.byteLength) {
+        stat = await this._writeStat(filepath, data.byteLength, { mode: stat ? stat.mode : 0o666 })
+        this.saveSuperblock() // debounced
+      }
+      if (encoding === "utf8") {
+        data = decode(data);
+      }
     }
+    if (!stat) throw new ENOENT(filepath)
     return data;
   }
   async writeFile(filepath, data, opts) {
@@ -1073,15 +1217,17 @@ module.exports = class PromisifiedFS {
       }
       data = encode(data);
     }
-    const stat = this._cache.writeFile(filepath, data, { mode });
+    const stat = await this._cache.writeStat(filepath, data.byteLength, { mode });
     await this._idb.writeFile(stat.ino, data)
     return null
   }
   async unlink(filepath, opts) {
     ;[filepath, opts] = cleanParams(filepath, opts);
-    const stat = this._cache.stat(filepath);
+    const stat = this._cache.lstat(filepath);
     this._cache.unlink(filepath);
-    await this._idb.unlink(stat.ino)
+    if (stat.type !== 'symlink') {
+      await this._idb.unlink(stat.ino)
+    }
     return null
   }
   async readdir(filepath, opts) {
@@ -1127,9 +1273,15 @@ module.exports = class PromisifiedFS {
     this._cache.symlink(target, filepath);
     return null;
   }
+  async backFile(filepath, opts) {
+    ;[filepath, opts] = cleanParams(filepath, opts);
+    let size = await this._http.sizeFile(filepath)
+    await this._writeStat(filepath, size, opts)
+    return null
+  }
 }
 
-},{"./CacheFS.js":4,"./HttpBackend.js":5,"./IdbBackend.js":6,"./Mutex.js":7,"./Stat.js":9,"./clock.js":10,"./errors.js":11,"./path.js":13,"isomorphic-textencoder":104,"just-debounce-it":106}],9:[function(require,module,exports){
+},{"./CacheFS.js":5,"./HttpBackend.js":6,"./IdbBackend.js":7,"./Mutex.js":8,"./Stat.js":10,"./clock.js":11,"./errors.js":12,"./path.js":14,"isomorphic-textencoder":104,"just-debounce-it":106}],10:[function(require,module,exports){
 module.exports = class Stat {
   constructor(stats) {
     this.type = stats.type;
@@ -1153,7 +1305,7 @@ module.exports = class Stat {
   }
 };
 
-},{}],10:[function(require,module,exports){
+},{}],11:[function(require,module,exports){
 module.exports = function clock(name) {
   performance.mark(`${name} start`);
   return function stopClock() {
@@ -1162,7 +1314,7 @@ module.exports = function clock(name) {
   };
 };
 
-},{}],11:[function(require,module,exports){
+},{}],12:[function(require,module,exports){
 function Err(name) {
   return class extends Error {
     constructor(...args) {
@@ -1179,11 +1331,12 @@ function Err(name) {
 
 const EEXIST = Err("EEXIST");
 const ENOENT = Err("ENOENT");
+const ENOTDIR = Err("ENOTDIR");
 const ENOTEMPTY = Err("ENOTEMPTY");
 
-module.exports = { EEXIST, ENOENT, ENOTEMPTY };
+module.exports = { EEXIST, ENOENT, ENOTDIR, ENOTEMPTY };
 
-},{}],12:[function(require,module,exports){
+},{}],13:[function(require,module,exports){
 const once = require("just-once");
 
 const PromisifiedFS = require('./PromisifiedFS');
@@ -1212,6 +1365,7 @@ module.exports = class FS {
     this.lstat = this.lstat.bind(this)
     this.readlink = this.readlink.bind(this)
     this.symlink = this.symlink.bind(this)
+    this.backFile = this.backFile.bind(this)
   }
   readFile(filepath, opts, cb) {
     const [resolve, reject] = wrapCallback(opts, cb);
@@ -1257,9 +1411,13 @@ module.exports = class FS {
     const [resolve, reject] = wrapCallback(cb);
     this.promises.symlink(target, filepath).then(resolve).catch(reject);
   }
+  backFile(filepath, opts, cb) {
+    const [resolve, reject] = wrapCallback(opts, cb);
+    this.promises.backFile(filepath, opts).then(resolve).catch(reject);
+  }
 }
 
-},{"./PromisifiedFS":8,"just-once":107}],13:[function(require,module,exports){
+},{"./PromisifiedFS":9,"just-once":107}],14:[function(require,module,exports){
 function normalizePath(path) {
   if (path.length === 0) {
     return ".";
@@ -1267,6 +1425,18 @@ function normalizePath(path) {
   let parts = splitPath(path);
   parts = parts.reduce(reducer, []);
   return joinPath(...parts);
+}
+
+function resolvePath(...paths) {
+  let result = '';
+  for (let path of paths) {
+    if (path.startsWith('/')) {
+      result = path;
+    } else {
+      result = normalizePath(joinPath(result, path));
+    }
+  }
+  return result;
 }
 
 function joinPath(...parts) {
@@ -1353,111 +1523,8 @@ module.exports = {
   split: splitPath,
   basename,
   dirname,
+  resolve: resolvePath,
 };
-
-},{}],14:[function(require,module,exports){
-'use strict';
-
-Object.defineProperty(exports, '__esModule', { value: true });
-
-class Store {
-    constructor(dbName = 'keyval-store', storeName = 'keyval') {
-        this.storeName = storeName;
-        this._dbName = dbName;
-        this._storeName = storeName;
-        this._init();
-    }
-    _init() {
-        if (this._dbp) {
-            return;
-        }
-        this._dbp = new Promise((resolve, reject) => {
-            const openreq = indexedDB.open(this._dbName, 1);
-            openreq.onerror = () => reject(openreq.error);
-            openreq.onsuccess = () => resolve(openreq.result);
-            // First time setup: create an empty object store
-            openreq.onupgradeneeded = () => {
-                openreq.result.createObjectStore(this._storeName);
-            };
-        });
-    }
-    _withIDBStore(type, callback) {
-        this._init();
-        return this._dbp.then(db => new Promise((resolve, reject) => {
-            const transaction = db.transaction(this.storeName, type);
-            transaction.oncomplete = () => resolve();
-            transaction.onabort = transaction.onerror = () => reject(transaction.error);
-            callback(transaction.objectStore(this.storeName));
-        }));
-    }
-    _close() {
-        this._init();
-        return this._dbp.then(db => {
-            db.close();
-            this._dbp = undefined;
-        });
-    }
-}
-let store;
-function getDefaultStore() {
-    if (!store)
-        store = new Store();
-    return store;
-}
-function get(key, store = getDefaultStore()) {
-    let req;
-    return store._withIDBStore('readonly', store => {
-        req = store.get(key);
-    }).then(() => req.result);
-}
-function set(key, value, store = getDefaultStore()) {
-    return store._withIDBStore('readwrite', store => {
-        store.put(value, key);
-    });
-}
-function update(key, updater, store = getDefaultStore()) {
-    return store._withIDBStore('readwrite', store => {
-        const req = store.get(key);
-        req.onsuccess = () => {
-            store.put(updater(req.result), key);
-        };
-    });
-}
-function del(key, store = getDefaultStore()) {
-    return store._withIDBStore('readwrite', store => {
-        store.delete(key);
-    });
-}
-function clear(store = getDefaultStore()) {
-    return store._withIDBStore('readwrite', store => {
-        store.clear();
-    });
-}
-function keys(store = getDefaultStore()) {
-    const keys = [];
-    return store._withIDBStore('readonly', store => {
-        // This would be store.getAllKeys(), but it isn't supported by Edge or Safari.
-        // And openKeyCursor isn't supported by Safari.
-        (store.openKeyCursor || store.openCursor).call(store).onsuccess = function () {
-            if (!this.result)
-                return;
-            keys.push(this.result.key);
-            this.result.continue();
-        };
-    }).then(() => keys);
-}
-function close(store = getDefaultStore()) {
-    return store._close();
-}
-
-exports.Store = Store;
-exports.get = get;
-exports.set = set;
-exports.update = update;
-exports.del = del;
-exports.clear = clear;
-exports.keys = keys;
-exports.close = close;
 
 },{}],15:[function(require,module,exports){
 function AbstractChainedBatch (db) {
@@ -8014,9 +8081,9 @@ exports.groupBy = groupBy;
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 let hiragana = "ぁあぃいぅうぇえぉおかがきぎくぐけげこごさざしじすずせぜそぞただちぢっつづてでとどなに" +
-    "ぬねのはばぱひびぴふぶぷへべぺほぼまみむめもゃやゅゆょよらりるれろゎわゐゑをんゔゕゖ";
+    "ぬねのはばぱひびぴふぶぷへべぺほぼぽまみむめもゃやゅゆょよらりるれろゎわゐゑをんゔゕゖ";
 let katakana = "ァアィイゥウェエォオカガキギクグケゲコゴサザシジスズセゼソゾタダチヂッツヅテデトドナニ" +
-    "ヌネノハバパヒビピフブプヘベペホボマミムメモャヤュユョヨラリルレロヮワヰヱヲンヴヵヶ";
+    "ヌネノハバパヒビピフブプヘベペホボポマミムメモャヤュユョヨラリルレロヮワヰヱヲンヴヵヶ";
 if (hiragana.length !== katakana.length) {
     throw new Error('Kana strings not same length?');
 }
@@ -11571,6 +11638,10 @@ function sync(gatty, lastSharedUid, uids, events, maxRetries = 3) {
                 return { newSharedUid, newEvents };
             }
             catch (pushed) {
+                {
+                    const log = yield git.log({ dir });
+                    console.error('Push failed', { pushed, log, staged });
+                }
                 // if push failed, roll back commit and retry, up to some maximum
                 const branch = (yield git.currentBranch({ dir })) || 'master';
                 yield gitReset({ pfs, git, dir, ref: 'HEAD~1', branch, hard: true, cached: false });
@@ -11611,7 +11682,7 @@ ${contents[idx]}`);
 }
 exports.inspect = inspect;
 
-},{"@isomorphic-git/lightning-fs":12,"filenamify":69,"isomorphic-git":88}],88:[function(require,module,exports){
+},{"@isomorphic-git/lightning-fs":13,"filenamify":69,"isomorphic-git":88}],88:[function(require,module,exports){
 (function (process,Buffer){
 'use strict';
